+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+## v1.3.0
+
+- Added: support for `Expo SDK` v50.
+
+## v1.2.0
+
+- Added: support for `react-native` v0.73.0.
+- Updated: `@svgr/core`, `@svgr/plugin-svgo`, and `@svgr/plugin-jsx` to v8.1.0.
+
+## v1.1.0
+
+- Updated: `@svgr/core` dependency to v8.0.0.
+- Updated: `@svgr/plugin-svgo` dependency to v8.0.1.
+- Added: `@svgr/plugin-jsx` dependency (required by SVGR).
+
 ## v1.0.0
 
 - Breaking: drop support for `react-native` versions older than 0.59